CkEditor options enabled - justify, showborders, pastefromword, pastetest, htmlwriter
Also enable forceSimpleAmbersand.

I took a look at 4.7 and there is enough change I think we can say '4.7 is different and if you are missing something when you get there we will need to look at it. These additions (requested by IMBA) don't seem to me to have a downside